cleanup flowchart e2e tests
